+++ featuretools/tests/primitive_tests/transform_primitive_tests/test_transform_primitive.py
@@ -986,7 +986,13 @@ class TestFileExtension(PrimitiveTestBase):
         transform, aggregation = find_applicable_primitives(self.primitive)
         primitive_instance = self.primitive()
         transform.append(primitive_instance)
-        valid_dfs(es, aggregation, transform, self.primitive.name.upper())
+        valid_dfs(
+            es,
+            aggregation,
+            transform,
+            self.primitive.name.upper(),
+            target_dataframe_name="sessions",
+        )
 
 
 class TestIsFirstWeekOfMonth(PrimitiveTestBase):