+++ docs/source/release_notes.rst
@@ -15,6 +15,7 @@ Future Release
     * Changes
         * Pinned pandas to 2.0.0 (:pr:`2533`)
         * Removed the ``Correlation`` and ``AutoCorrelation`` primitive as these could lead to data leakage (:pr:`2537`)
+        * Remove IntegerNullable support for ``Kurtosis`` primitive  (:pr:`2537`)
     * Documentation Changes
     * Testing Changes