+++ docs/source/release_notes.rst
@@ -13,7 +13,7 @@ Future Release
             ``IsUnique`` (:pr:`2537`)
     * Fixes
     * Changes
-        * Pinned pandas to 2.0.0 (:pr:`2533`)
+        * Pinned pandas to < 2.0.0 (:pr:`2533`)
         * Removed the ``Correlation`` and ``AutoCorrelation`` primitive as these could lead to data leakage (:pr:`2537`)
         * Remove IntegerNullable support for ``Kurtosis`` primitive  (:pr:`2537`)
     * Documentation Changes